Fizz: Add GPIOs used for CEC
The Fizz hardware has three pins for CEC. One GPIO is used
as a pull-up. It is always an output and always high. The
second GPIO is a data output. The third GPIO is the data
input that can also be configured as a timer input (TA1).

Signed-off-by: Stefan Adolfsson <sadolfsson@chromium.org>

BUG=b:76467407
BRANCH=none
TEST=Read and write the newly added GPIOs using ectool
CQ-DEPEND=CL:1030218

Change-Id: Ia33b36a0cdaa40fd1a4f7aa66a092b5833bf5cf8