board/yorp: fix PMIC_EN gpio
Apparently the symbol in the yorp schematic is incorrect for our part.
The PMIC_EN signal on ball H6 is actually GPIO72 -- not GPIOD7. Adjust
the gpio used for PMIC_EN. Additionally, GPIO72 needs to be put in
gpio mode since it defaults to PWRGD functionality. Do this in
board_init() only when we haven't jumped to the image -- power-on-reset.

BUG=b:78352179
TEST=Built. Booted. PMIC_EN goes up and down as expected.
BRANCH=none

Change-Id: I955f9a24e0fbecb0cda1380c237fa44c9a575e45
Signed-off-by: Aaron Durbin <adurbin@chromium.org>