cheza: Support confirmation of power lost
Add a handler to save the deferred power signals after a delay of
interrupt. Confirm the power lost if the POWER_GOOD signal keeps
low for a while (both the current and the deferred signals are low).

BRANCH=none
BUG=b:78455067
TEST=Toogle EC GPIO SYS_RST_L for a low pulse to execute PMIC reset
sequence and verified AP reset but not a transition S0 -> S5.
TEST=Toogle EC GPIO PMIC_KPD_PWR_ODL for a low pulse to execute PMIC
shutdown sequence and verified a power-lost transition S0 -> S5.

Change-Id: I8ed789d701e834195865bfdf2d302388d42618d2
Signed-off-by: Tom Wai-Hong Tam <waihong@google.com>