yorp: lower i2c speeds to 100kHz
We need to tune the i2c parameters for the nuvoton chip on our yorp board
to be able to operate at 400kHz. Currently we do not need the extra speed
or bandwidth, so we are reverting to a lower speed where the default
timing parameters work well.

BRANCH=none
BUG=b:78554726,b:78225299
TEST=HDMI over TypeC works on yorp

This reverts commit 2e7e6665b1e712a950eef8ae3d3f64ae0f1d2ec1.

Change-Id: Ic4ba1d5ef25661bd6c7f9490450af65b4e1393ad
Signed-off-by: Jett Rink <jettrink@chromium.org>