Fizz: Start TCPC reset pin as GPIO_INPUT
The GPIO pin connected to PS8751's reset pin is configured as GPIO_OUT_LOW.
This causes the EC to generate a reset pulse when Cr50 resets the EC on
power-on reset. When PS8751 is reset before HP Z32 enters PD negotiation,
HP Z32 drops VBUS thus the system browns out.

This patch sets the pin as GPIO_INPUT first. This is reconfigured to
GPIO_LOW_OUT in HOOK_INIT. This fix works as long as the H1 reset happens
before HOOK_INIT.

BUG=b:74354232
BRANCH=none
TEST=Verify Fizz boots on HP Z32.

Change-Id: Ieae7c25e8261cf3b708e76c8340b07af5a61720f