Nami: Start TCPC reset pins as GPIO_INPUT
The GPIO pins connected to TCPC reset pins are configured as
GPIO_ODR_HIGH and GPIO_OUT_LOW. This causes the EC to generate a reset
pulse when Cr50 resets the EC on power-on reset. When a TCPC is reset
before HP Z32 enters PD negotiation, HP Z32 does not show a picture.

This patch sets the pin as GPIO_INPUT first. This is reconfigured to
GPIO_ODR_HIGH and GPIO_LOW_OUT in HOOK_INIT.

BUG=b:110441862
BRANCH=none
TEST=Verify Sona shows picture on HP Z32 after:
1. Cold reset (by servo)
2. Power off then on

Change-Id: Ieae7c25e8261cf3b708e76c8340b07af5a61720f