cr50: drop generating .hex file and add size calculations
Room remaining if flash is an ever important property of the Cr50
images. This patch adds a makefile rule with will calculate how much
free room is left in flash with the current image.

The calculation uses the '__hey_flash_used' value calculated by the
linker, which is the sum of text, RO data and initialized data segment
sizes, and some constants of Cr50 memory layout RO and NVMEM
segments).

BRANCH=cr50, cr50-mp
BUG=b:65253310
TEST=observed reported values:
    $ make BOARD=cr50 -j
    ...
    *** 13272 bytes still available in flash ****
    $ rm -rf build/cr50/
    $ make BOARD=cr50 CR50_DEV=1 -j
    *** 9872 bytes still available in flash ****

Change-Id: I66e23dc72905988b21666c5dc9608d92e3fead50
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>