grunt: Increase CONFIG_EXTPOWER_DEBOUNCE_MS to 200
When plugging in USB-C power with no battery (or cutoff battery),
the ACOK signal from the ISL9238 charger sometimes has a negative
pulse of up to 175 ms after VBUS increases from 5V to 20V.
The width of the pulse varies from board to board.
Increase CONFIG_EXTPOWER_DEBOUNCE_MS to 200 ms (from default of 30)
to ignore this negative pulse.
If we think we are running with no battery and no AC then we will
set the charger voltage to 2.048V, and therefore lose power.

BUG=b:77455171
BRANCH=none
TEST=Plug power into Grunt with no battery.

Change-Id: I9c0f358985e4c630daef93d61c49b87c2e11a480
Signed-off-by: Edward Hill <ecgh@chromium.org>