usb_pd_protocol: Add PD_ROLE_DISCONNECTED for data role swaps
Add a state to indicate that a data role is actually disconnected and
notify the board-level data swap function with this state when a cable
is unplugged.  This allows the board to clean up and restore any state
that may have been set up with a data role swap.

BUG=b:78308749
BRANCH=eve,poppy
TEST=manual on eve: plug in C-to-C cable, execute data swap on port 0
with 'pd 0 swap data' and ensure the OTG pins are asserted.  Then unplug
the cable and ensure OTG pins are now deasserted.

Change-Id: I41f859c74c4c1ce2523246250ab324c5f4177099
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id:
Original-Change-Id: I7d8fff22dd5836b4b5af54f0ede71ee1b6e40b5c
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>