Update patch set 5
Patch Set 5:

Same here (as the parent cl), the current problem is that we have a quirk in the signal we chose to use as a power good. It's a property of the chipset, it should probably go in power/

Unless you think we shouldn't use that signal as a power good to begin with and we'll fix it in the next hardware revision.

Patch-set: 5
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>