atlas: config PROCHOT GPIO as input
we need to configure EC_PROCHOT_ODL as an input because the EC isn't
driving it correctly.  we changed the polarity of EC_PROCHOT for atlas
and similar boards, but the EC codebase has this hard-coded as
active-high.

this is a short-term fix until we implement a more general PROCHOT
"polarity" feature.

BUG=b:78911901,b:79266467
BRANCH=none
TEST=checked voltage drop across in-line resistor on EC_PROCHOT and
AP can now run above 400MHz

Change-Id: I8c3224c62ea7af4f386062d39c248d418e73fa53
Signed-off-by: Caveh Jalali <caveh@chromium.org>