charge_manager: return valid data for BJ when not connected
When we boot the system from a USB PD port and keep the dedicated
port not connected, we get incorrect information about the dedicated
port when we call charge_manager_fill_power_info.
The function returns the max voltage and max current for USB PD ports,
instead of the value for the BJ, and the charger type is set to
None, which also has the undesirable effect of preventing the kernel
to effectively recognizing this port as dedicated during driver's probe
when the BJ is not connected.

The commit fix this issue by always returning dedicated port as
dedicated, and setting the voltage_max and current max to 0.

BRANCH=None
BUG=chromium:841944
TEST=Checked that the values returned for the dedicated port are the
one expected.

Change-Id: I4d35ba0691cdd681643918e89ec50e8e6d7be8bb
Signed-off-by: Fabien Parent <fparent@baylibre.com>