HACK: charge_manager: find correct supply when PWR > 0 and CUR = 0
On Fizz if we boot from type-C and later on plug the BJ,
the BJ available voltage is set to a non-zero value but the current
is set to 0. This creates two issues:
    * the power source is never switched to BJ because PWR = 0
    * charge_manager_fill_power_info will return that the BJ is
      unplugged because it won't find the supply since PWR = 0

This HACK is only aiming to workaround the second point. When looking
for the supply associated with the port, we won't look for supply
where the PWR > 0, but where supply where the voltage > 0.
With this HACK we can now get the real information with
charge_manager_fill_power_info about if a supply is plugged, and
what its voltage is.

Change-Id: I7342829aa3429102d8452e9d2beba38d1701027f
Signed-off-by: Fabien Parent <fparent@baylibre.com>