cheza: Enable PP5000 after initializing adapter current limit
Refine the previous hack.

The adapter current limit is set at the beginning of the charger
task. Enabling PP5000_A should be done after that. So do it in
board_post_init(), a deferred function, in the context of the hook
task which has a lower priority than the charger task.

BRANCH=none
BUG=b:79353631
TEST=On battery plugged and unplugged cases, checked the device can
source VBUS to USB port-0 and port-1.

Change-Id: Ie9b8dff6e10d97dffd554b382595e5e7a70875e6
Signed-off-by: Wai-Hong Tam <waihong@google.com>