cheza: Enable PP5000 on G3->S5 and change the initial input current
Remove the previous hack of force increasing the adapter current.

In the charger init hook, it sets the input current to a board
config value. Increase this value to 1500mV which can supply enough
power to the system even no battery.

The PP5000 rail is now turned on/off during the chipset power
sequence G3 <-> S5.

The init hook is executed before the chipset task. It guarantees
the charger is configured (the initial input current) before we
enable the PP5000 rail.

It also makes sure it has enough power to enable the 5V rail. If
the battery is in low level or unplugged and the charger adapter
doesn't supply enough power, don't enable the 5V rail and don't
boot into S5.

BRANCH=none
BUG=b:79353631
TEST=On battery plugged and unplugged cases, checked the device can
source VBUS to USB port-0 and port-1.
TEST=Unplug battery and use a low-power adapter, can't boot up AP.
See the "Not enough power to boot" message instead of a reboot loop.

Change-Id: Ie9b8dff6e10d97dffd554b382595e5e7a70875e6
Signed-off-by: Wai-Hong Tam <waihong@google.com>