cheza: Enable PP5000 after initializing adapter current limit
Refine the previous hack.

If no battery plugged, the only power source is the adapter. But the
initial adapter current limit 1500mA is not enough for the whole
system. Initialize it to the max the board can support.

After PD negotiation, it will change to a proper value.

BRANCH=none
BUG=b:79353631
TEST=On battery plugged and unplugged cases, checked the device can
source VBUS to USB port-0 and port-1.

Change-Id: Ie9b8dff6e10d97dffd554b382595e5e7a70875e6
Signed-off-by: Wai-Hong Tam <waihong@google.com>