WIP/HACK: Compress bootblock using LZ4
This compresses each of the 512-bytes eMMC blocks individually,
which uses the least amount of SRAM on EC, but is surely not
optimal.

Also, this only decompresses the blocks after DMA is completed:
in theory we should be able to decompress in parallel (eMMC clock
is so fast though, that the gains may not be very important.

Needs cleanup, and figuring out if this is the desired approach.

[3.499490 transfer]
[3.521406 goIdle]
Transfer time: 22ms (vs 13ms when no compression is used)
7.5ms of these are spend decompressing the data

Change-Id: I0a7746680052d84aa19886da45597e0a4ec5acc7
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>