Update patch set 6
Patch Set 6:

(3 comments)

Per our conversation, this solution guarantees that lower priority tasks have gotten the chance to start performing a logical block of TCPC interaction (due to the interrupt), but it does not guarantee that the logical block in a task is done.

I think we can add some safe guards to the task themselves that use and extend the ideas in this CL. Let's talk on Monday about it.

Patch-set: 6
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>