Update patch set 1
Patch Set 1:

Hi Randall, I spent most of today looking at the CLs that would be required to enable the new UHEPI for EC-AP communication, the most important of which I believe are these two:

* CL:576047 for EC side
* CL:871030 for coreboot side

However, it turns out that the eve firmware branch has diverged quite a bit, and merging hell kept getting deeper and deeper.  UHEPI is quite a big change and involves many different CLs.

This CL still makes use of the old API but patches on support for 64-bit responses through a one-line change, which I think is quite reasonable provided we document it properly.  If in the future we decide to brave merging hell and backport UHEPI, removing this CL would be a simple revert.

Does this sound reasonable?

Patch-set: 1
Reviewer: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>