nocturne: Fix EC hibernate.
This commit adds the appropriate hibernate flags to the hibernate wake
pins.  It additionally, adds a board specific hibernate function which
sets up the PSL pins for wake as well as writing to the ROP PMIC to
disable all the power rails.

BUG=b:79713379
BRANCH=master
TEST=Enter `hibernate` on EC console, verify that system can wake from
AC insertion, power button press, and lid switch.

Change-Id: I16e0e641ed3a7daaecd2661a9525d6653cbb1dc4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>