cheza: Make LED work even AP is in G3
When AP is in G3, the PP5000 rail is disabled so the LEDs don't
work. Enable the PP5000 for the LEDs.

If the battery/adapter can't supply enough power, don't enable it.

The power source of LED will be moved to an always-rail and not
need this patch.

BRANCH=none
BUG=b:79749048
TEST=AP in G3, tested several scenarios: charge, low-level battery,
charge but no battery.

Change-Id: Iad452b3e313b0ef80348a8a799181c699f9b1a05
Signed-off-by: Wai-Hong Tam <waihong@google.com>