Update patch set 3
Patch Set 3:

Jett, thanks for the clarification with the new CL. Seems like the CONFIG_TABLET_SWITCH got me confused with the CONFIG_TABLET_MODE_SWITCH.

Btw, from the HGDE sensor datasheet tablet mode is active low although the GPIO name in our schematics does not reflect the same. I see that the gpioget is always high which is showing inactive at this point, since the chassis lid is not in tablet mode.

Patch-set: 3