isl9238: enable discharged battery detection
we have a somehwat obscure case where a discharged battery will not
respond over I2C and we don't properly handle the battery until we
have read some information from it.

we need to set the ChargeCurrentLimit register to a non-zero value so
the charger will power the battery to enable its I2C controller.  the
actual current limit is defined by register "Control2<15:14>" which
we leave at the default 256mA.

BUG=b:79354967
BRANCH=none

Change-Id: Id68e9218c17ac137f166c3869c8d74c97abcddae
Signed-off-by: Caveh Jalali <caveh@chromium.org>