make get_interrupt_context() static inline
It's only used in the same file and should always be inlined. Should
save a couple of bytes because the separate function isn't necessary
anymore in the object file and also de-confuses gcc 8.1 that seems to
not inline it (since inline is just a hint) but then drops the function
(presumably because it's small, marked inline, and comes with no
prototype).

BUG=b:65441143
BRANCH=none
TEST=builds with gcc 8.1

Change-Id: I881a5b9f13192dd11748d8a3060788f95a84dec0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>