Update patch set 1
Patch Set 1:

as it is now, sleepmask command allows to change the mask. This ability should be only in images compiled with CR50_DEV, otherwise we just want to see it, but don't want to be able to modify.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>