sensors: sync: Add the extra missed syncs in the protocol
In case the EC was busy and didn't have time to process motion_sense task
events, we might miss the recording of a sync timestamp:

0 ms, sync event 0
33 ms, sync event 1
66 ms, interrupt came, but we didn't have time to record it in the task
99 ms, sync event 3

Note the missing sync event 2.
But with this change, we add an entry for sync event 2 in our buffer:

0 ms, sync event 0
33 ms, sync event 1
99 ms, sync event 2
99 ms, sync event 3

Note while the sync event 2 timestamp is wrongly set to 99 ms (instead of
66 ms), it matches what the bmi160 driver does too. The AP already knowns of
this case and does spreading of the timestamps (by interpolation it'll come
up with 66ms).

BRANCH=master
BUG=b:73551961, b:67743747
TEST=Change command_sync to call 2 interrupts instead of just one, that
should emulate the "motion_sense task is busy" condition. Then on the AP
you should see no skipped sync events (look at the counter value):

0 34 -190 8179 3000260898305
2 20 0 0 3000340020702
0 44 -176 8176 3000342039685
0 41 -150 8204 3000423183036
2 21 0 0 3000423927283
2 22 0 0 3000423927283
0 42 -138 8268 3000504338186

instead of:

0 34 -190 8179 3000260898305
2 20 0 0 3000340020702
0 44 -176 8176 3000342039685
0 41 -150 8204 3000423183036
2 22 0 0 3000423927283
0 42 -138 8268 3000504338186

Change-Id: I9fcb1fb8b35eb5f8ffcc21afbfcb0f0d9bc33804
Signed-off-by: Alexandru M Stan <amstan@chromium.org>