atlas: improve discharged battery handling
we normally try to find out a few things about a battery (like charge
level) before actaully applying charging power to it.  when the
battery is completely discharged, the controller on the battery can't
respond as it is not self-powered.  so, we have to avoid all
operations that depend on the battery responding in the battery
discovery/initialization path.

as long as we report that a battery is present and it is not
responsive, the charger task will enter ST_PRECHARGE which means it'll
provide a "precharge" current to the battery to try to talk to it.
this allows the battery's controller to report battery parameters
allowing our charger task can do the right thing.

BUG=b:79354967
BRANCH=none
TEST=atlas now discovers the discharged battery reliably

Change-Id: I5e5a3abda07508eb791b712fb2f9b9f5fe383e07
Signed-off-by: Caveh Jalali <caveh@chromium.org>