atlas: improve discharged battery handling
we normally try to find a few things about a battery (like charge
level) before we actaully apply charging power to it.  when the
battery is completely discharged, the controller on the battery can't
respond as it is not self-powered.  so, we have to avoid all
operations that depend on the battery responding in the battery
discovery/initialization path.

also, we were blindly applying a thermal throttle to charging based on
unavailable battery temperature data.

finally, once the battery is responsive using the "precharge current",
it briefly requests requests (0 vols, 0 amps) - presumably while its
controller is trying to figure out what's going on...  we respect this
and stop charging, but that's probably a really bad idea since the
battery has had very little chance accept enough charge to self-power
its controller.  enabling "REQUESTS_NIL_WHEN_DEAD" gets around that.

BUG=b:79354967
BRANCH=none
TEST=atlas now discovers the discharged battery reliably

Change-Id: I5e5a3abda07508eb791b712fb2f9b9f5fe383e07
Signed-off-by: Caveh Jalali <caveh@chromium.org>