flash_ec: npcx_uut: If the DUT is unpowered, don't restore settings
Normally, the flash_ec script first saves some original servo settings,
like ec_boot_mode, etc. and then restores it back at the end. If the
DUT is not powered, the servo pin is open and servo think the
ec_boot_mode is "on" (the cold_reset is "on" too). When the flash_ec
script fails and then restores, the ec_boot_mode is set to a real "on".
So people will see the EC UART not work as EC is in a flashing mode.

This patch tries to prevent this issue.

It checks whether the DUT is unpowered, through the servo control
cold_reset to indicate. If yes, the servo settings are actually not
correct and we should exit without restore. Since this check is in a
very early state, no servo setting is changed that time; it is fine
to skip restoring the servo settings.

BRANCH=none
BUG=b:80305869
TEST=Ran the flash_ec script when the DUT is unpowered; and then reran
again when powered. It worked fine and EC UART showed-up afterward.

Change-Id: I127eb1dfc19a5f4b214eaadacaf25bcb1e3c0395
Signed-off-by: Wai-Hong Tam <waihong@google.com>