atlas: ignore unavailable battery temp readings
the battery temperature field is only valid after we've actually
managed to read the battery temperatore parameter.  so, if the temp
field is marked "BAD", don't even look at it.

this addresses a case where we were removing charge current during the
precharge phase - basically removing charge current from a battery
that we're trying to power so we can talk to its I2C controller.

BUG=b:79354967
BRANCH=none
TEST=instrumented code to verify we don't request 0 amps in ST_PRECHARGE

Change-Id: I3b40903506fa949c14ecaf577f134f31cfcf8fb7
Signed-off-by: Caveh Jalali <caveh@chromium.org>