Update patch set 1
Patch Set 1:

I feel that this is only covering up the symptom.

Besides, I wouldn't want to use cold_reset asserted as an indicator the board was off.

Isn't the issue here that we're restoring some bad settings? Why is the state of ec in programming mode saved as a the real state we should be restoring?

Patch-set: 1
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>