Update patch set 1
Patch Set 1:

Normally, the flash_ec script first saves some original servo settings, like ec_boot_mode, etc. and then restores it back at the end. If the board doesn't have power, the servo pin is open and servo think the ec_boot_mode is "on" (the cold_reset is "on" too). When the flash_ec script fails (the board not powered) and then restores, the ec_boot_mode is set to a real "on". So people will see the EC UART not work as EC is in a flashing mode.

This patch to prevent the above issue.

It checks whether the board is unpowered or not. If yes, the servo settings are actually not correct and we should exit. Since this check is in a very early state, no servo setting is changed that time; it just skips the exit hook which always restores settings.

I am not sure what is the best indication of the board powered or not. Any suggestion?

Patch-set: 1
Reviewer: Gerrit User 1000407 <1000407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>