Update patch set 1
Patch Set 1:

So you're saying we're reading ec_boot_mode when saving the state, as in the actual pin. And because the pin is high usually, but low when the board is off we save the wrong state.

Could we change ec_boot_mode to not actually read the pin state? I know we're using is as an open drain (so we can OR with other things that want to drive this), but that doesn't mean reading it should actually read the actual signal, we should read the state we're outputting instead.

Patch-set: 1
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>