Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Didn't CONFIG_BRINGUP cover this, by changing PWRBTN_STATE_INIT_ON to PWRBTN_STATE_IDLE?

That would have worked too. This does cover the scenario for later on in the cycle though since the ITE EC is powered from the servo and will try to boot the system incorrectly without this change.

I never defined the CONFIG_BRINGUP option. That seems helpful :)

Patch-set: 5