ppc: remove EC ability to drive PPC GPIO
We figured out that having the EC drive the EN_SNK input of a PPC will
lead to brownout or lockout conditions since the EC will float the GPIO
when the EC reset. When the PPC input floats, the power path to the rest
of the board is cut.

If you mix this scenario with a PPC that hard-codes the over voltage
protect of vbus at 6.8V in dead battery mode, you can lock the
previously negotiated 20V Vbus out of the system until you physically
remove the USBC connection and reinsert it.

BRANCH=none
BUG=b:78896495,b:78021059
TEST=removing unused code. everything still works.

Change-Id: Ie9d51af766bc08e7d9b46b4610c0f84d3d0961f8
Signed-off-by: Jett Rink <jettrink@chromium.org>