Update patch set 1
Patch Set 1:

> Can you program the board while it's unpowered?

No.

UUT is similar to the STM32 programming method (including the pin (BOOT0 on stm32) that you hold low to get in the programming mode). Some maskrom on the EC runs and talks uart to a programmer, the EC essentially self programs based on that uart protocol. Power has to be on.

> What's the sequence here?
save state, assert pin, flash, deassert pin(by restoring state)

The issue is that if we save state with the device unpowered (honest developer mistake), save state incorrectly saves the boot pin as asserted. Then we restore the state (after the flash fails, but the fact that it failed is a red herring) and make the pin asserted (even though it should be deasserted).

I still think it makes no sense to read the state like that. Just because something else (in this case the lack of power) is asserting a signal doesn't mean we should start asserting it too when we restore the state.

write_pin(read_pin(pin_that_is_open_drain)) is destructive. It's not just a noop. Let's not do this.

Patch-set: 1
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>