chip/stm32/clock: Initialize 'alarm_us'
The new GCC build shows a warning/error in clock-f.c:

'alarm_us' may be used uninitialized in this function
[-Werror=maybe-uninitialized]

This is actually a fake warning. In the context of the logic,
there is no way 'alarm_us' would be used uninitialized.

But let's still initialize 'alarm_us' to clear the compiler warning.

BUG=none
BRANCH=scarlet
TEST='USE=coreboot-sdk emerge-scarlet chromeos-ec'
TEST=make buildall -j

Change-Id: I7a0642cbe03c5a0adb6997ddc80c9cb797715749
Signed-off-by: Philip Chen <philipchen@google.com>