Update patch set 1
Patch Set 1:

(2 comments)

Yes, that sounds reasonable.

Please use a different bug#, though, so it's clear this is a hack for campfire and not part of the actual fix for the bug.

Though if it's a campfire-only fix, it seems like you could have just reused bit 31 for your AltOS event and not need to change the struct size at all.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>