console: Do not flush the console in console_init
Currently, console_init calls cflush() twice, once before
"Console is enabled" string is printed, once afterwards.

The reason is that firmware_ECBootTime looks for that string,
and it may get corrupted/interleaved with others if the EC
is busy during initialization.

The problem here is that the CONSOLE task may have higher
priority than other tasks (for good reasons), but, on boot,
there are other more critical tasks that need to run (e.g.
RW image verification), rather than busy-looping waiting for
the console to be flushed.

 1. We remove the cflush() after the ccprintf: Once the
    string has be appended to the TX buffer, it is
    guaranteed to be printed in sequence (unless the TX buffer
    overflows, but let's assume this won't happen...)
 2. The cflush before the ccprintf actually serves the purpose
    of waiting for the EC to be less busy, so that no other
    task will preempt CONSOLE task. We just replace it with
    a msleep(300) that will let the EC run other tasks.
    Developers do not care about being able to type on the
    console before that point anyway, as the console is very
    busy and unusuable anyway.

We should consider modifying firmware_ECBootTime to not rely
on the string (or at least look for it in a more lazy way,
allowing for gaps in the string).

BRANCH=poppy
BUG=b:35647963
BUG=chromium:687228
TEST=Flash staff, see that RW verification starts at 0.000952
     instead of 0.028093 (=> 27 ms faster).

Change-Id: I794e48eb69cc647c4595fd80265adee4a434d566
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>